test: isolate standalone changed-run coverage
